Move AmazingRace35 to archive as suggested in issue comment
